# docs/openapi/v0.bundled.yaml
openapi: 3.1.0
info:
  title: Plataforma GALA API
  version: 0.1.0
  description: |
    Plataforma GALA (Generación Audiovisual Local con Avatares). Contrato API v0 (local, sin auth).
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /templates:
    post:
      tags:
        - Templates
      summary: Create template
      operationId: createTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags:
        - Templates
      summary: List templates
      operationId: listTemplates
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplatesListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    templateById:
      get:
        tags:
          - Templates
        summary: Get template
        operationId: getTemplate
        parameters:
          - $ref: '#/components/parameters/templateId'
        responses:
          '200':
            description: OK
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/TemplateResponse'
          '404':
            $ref: '#/components/responses/NotFound'
          '500':
            $ref: '#/components/responses/InternalError'
      patch:
        tags:
          - Templates
        summary: Update template
        operationId: updateTemplate
        parameters:
          - $ref: '#/components/parameters/templateId'
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateTemplateRequest'
        responses:
          '200':
            description: Updated
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/TemplateResponse'
          '400':
            $ref: '#/components/responses/ValidationError'
          '404':
            $ref: '#/components/responses/NotFound'
          '409':
            $ref: '#/components/responses/Conflict'
          '500':
            $ref: '#/components/responses/InternalError'
      delete:
        tags:
          - Templates
        summary: Delete template
        operationId: deleteTemplate
        parameters:
          - $ref: '#/components/parameters/templateId'
        responses:
          '204':
            description: Deleted
          '404':
            $ref: '#/components/responses/NotFound'
          '500':
            $ref: '#/components/responses/InternalError'
  /templates/{templateId}:
    get:
      tags:
        - Templates
      summary: Get template
      operationId: getTemplate
      parameters:
        - $ref: '#/components/parameters/templateId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags:
        - Templates
      summary: Update template
      operationId: updateTemplate
      parameters:
        - $ref: '#/components/parameters/templateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTemplateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - Templates
      summary: Delete template
      operationId: deleteTemplate
      parameters:
        - $ref: '#/components/parameters/templateId'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /jobs:
    post:
      tags:
        - Jobs
      summary: Create render job (batch)
      operationId: createJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags:
        - Jobs
      summary: List jobs
      operationId: listJobs
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - QUEUED
              - RUNNING
              - DONE
              - FAILED
        - in: query
          name: template_id
          schema:
            type: string
        - in: query
          name: model_id
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobsListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
...
    templateId:
      name: templateId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
            - code
            - message
      required:
        - error
...
    TemplateFormat:
      type: object
      properties:
        width:
          type: integer
        height:
          type: integer
        fps:
          type: integer
      required:
        - width
        - height
        - fps
    Template:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        name:
          type: string
        duration_ms:
          type: integer
          nullable: true
        format:
          $ref: '#/components/schemas/TemplateFormat'
        params_schema:
          type: object
          additionalProperties: true
        defaults:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - type
        - name
        - created_at
    TemplatesListResponse:
      type: object
      properties:
        templates:
          type: array
          items:
            $ref: '#/components/schemas/Template'
      required:
        - templates
    CreateTemplateRequest:
      type: object
      required:
        - type
        - name
      properties:
        type:
          type: string
        name:
          type: string
        duration_ms:
          type: integer
          nullable: true
        format:
          $ref: '#/components/schemas/TemplateFormat'
        params_schema:
          type: object
          additionalProperties: true
        defaults:
          type: object
          additionalProperties: true
    TemplateResponse:
      type: object
      properties:
        template:
          $ref: '#/components/schemas/Template'
      required:
        - template
    UpdateTemplateRequest:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        duration_ms:
          type: integer
          nullable: true
        format:
          $ref: '#/components/schemas/TemplateFormat'
        params_schema:
          type: object
          additionalProperties: true
        defaults:
          type: object
          additionalProperties: true
...
    Job:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - QUEUED
            - RUNNING
            - DONE
            - FAILED
        template_id:
          type: string
        model_ids:
          type: array
          items:
            type: string
        params:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - status
        - created_at
    JobsListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
      required:
        - jobs
    CreateJobRequest:
      type: object
      required:
        - template_id
        - model_ids
      properties:
        name:
          type: string
        template_id:
          type: string
        model_ids:
          type: array
          items:
            type: string
        inputs:
          type: object
          additionalProperties: true
        params:
          type: object
          additionalProperties: true
        output:
          type: object
          additionalProperties: true
    JobResponse:
      type: object
      properties:
        job:
          $ref: '#/components/schemas/Job'
      required:
        - job
...